1. 完善校准.

+++ app/src/main/java/alauncher/cn/measuringinstrument/view/activity_view/CalibrationActivityView.java
@@ -7,5 +7,5 @@ public interface CalibrationActivityView {
 
     void onUIUpdate(CalibrationBean bean);
 
-    void onDataUpdate(double[] values);
+    void onDataUpdate(int[] values);
 }
